# backend/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SmartRetail API",
    "description": "SmartRetail Backend API – cleaned & valid OpenAPI 3",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://skinanalysis.life",
      "description": "Production"
    },
    {
      "url": "http://localhost:9999",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/taxs/preview": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "Authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "x-store-id",
            "in": "header",
            "type": "string"
          },
          {
            "name": "req",
            "in": "query",
            "type": "string"
          },
          {
            "name": "periodKey",
            "in": "query",
            "type": "string"
          },
          {
            "name": "periodType",
            "in": "query",
            "type": "string"
          },
          {
            "name": "storeId",
            "in": "query",
            "type": "string"
          },
          {
            "name": "monthFrom",
            "in": "query",
            "type": "string"
          },
          {
            "name": "monthTo",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/taxs/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "Authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "x-store-id",
            "in": "header",
            "type": "string"
          },
          {
            "name": "req",
            "in": "query",
            "type": "string"
          },
          {
            "name": "storeId",
            "in": "query",
            "type": "string"
          },
          {
            "name": "periodType",
            "in": "query",
            "type": "string"
          },
          {
            "name": "periodKey",
            "in": "query",
            "type": "string"
          },
          {
            "name": "status",
            "in": "query",
            "type": "string"
          },
          {
            "name": "isClone",
            "in": "query",
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "Authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "x-store-id",
            "in": "header",
            "type": "string"
          },
          {
            "name": "req",
            "in": "query",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "storeId": {
                  "example": "any"
                },
                "req": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/taxs/{id}/clone": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "Authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "x-store-id",
            "in": "header",
            "type": "string"
          },
          {
            "name": "req",
            "in": "query",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "storeId": {
                  "example": "any"
                },
                "req": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/taxs/{id}/approve": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "Authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "x-store-id",
            "in": "header",
            "type": "string"
          },
          {
            "name": "req",
            "in": "query",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "storeId": {
                  "example": "any"
                },
                "req": {
                  "example": "any"
                },
                "action": {
                  "example": "any"
                },
                "rejectionReason": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/taxs/{id}/export": {
      "get": {
        "description": "",
        "produces": ["application/pdf; charset=utf-8"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "Authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "x-store-id",
            "in": "header",
            "type": "string"
          },
          {
            "name": "req",
            "in": "query",
            "type": "string"
          },
          {
            "name": "format",
            "in": "query",
            "type": "string"
          },
          {
            "name": "Content-Type",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/taxs/{id}": {
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "Authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "x-store-id",
            "in": "header",
            "type": "string"
          },
          {
            "name": "req",
            "in": "query",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "storeId": {
                  "example": "any"
                },
                "req": {
                  "example": "any"
                },
                "declaredRevenue": {
                  "example": "any"
                },
                "taxRates": {
                  "example": "any"
                },
                "revenueByCategory": {
                  "example": "any"
                },
                "specialConsumptionTax": {
                  "example": "any"
                },
                "environmentalTax": {
                  "example": "any"
                },
                "notes": {
                  "example": "any"
                },
                "internalNotes": {
                  "example": "any"
                },
                "status": {
                  "example": "any"
                },
                "isFirstTime": {
                  "example": "any"
                },
                "supplementNumber": {
                  "example": "any"
                },
                "taxpayerInfo": {
                  "example": "any"
                },
                "otherFields": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "Authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "x-store-id",
            "in": "header",
            "type": "string"
          },
          {
            "name": "req",
            "in": "query",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "storeId": {
                  "example": "any"
                },
                "req": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "Authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "x-store-id",
            "in": "header",
            "type": "string"
          },
          {
            "name": "req",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}